Replace die() with panic()
This removes die() altogether.

Additionally panic when the key cannot be inserted. We see this a lot:

    Error inserting key 03aff758af79d0338ce6ecf463e36b38 into database: exec errors: &{"" {""} [{"Neo.ClientError.Transaction.ConcurrentRequest" "The requested transaction is being used concurrently by another request."}] ""}

But we don't get a stacktrace. Part of the reason is that the code
explicitly checks for a Postgres error and only exits in those cases.
Here, just replace the whole thing with a panic(). We'll sort it out.